EntityManager handles references between tokens and entities via a Map. Whenever the boundaries of an entity are changed the EntityManager adjusts the mentioned Map. This Map allows an easy access to the entities that reference a certain token. Additionally, due to the access by token index, it is easier to convert a state into a String representation.

+++ src/Corpus/Token.java
@@ -3,7 +3,7 @@ package Corpus;
 
 public class Token {
 
-	long index;
+	int index;
 	int start, stop;
 	String text;
 
@@ -19,7 +19,7 @@ public class Token {
 		return text;
 	}
 
-	public long getIndex() {
+	public int getIndex() {
 		return index;
 	}
